Corrects init definition, imports, and doc-strings
The default value of infinity for the conductor conductivity is now
properly defined using float('inf').

No need to import Media because Distributed circuit has the to_meters
function. The function call is updated to reflect this hierarchy.

The doc string was too long so transmission line was shortened to t-line